feat(config): Add type for templated password command
I use the password-command flag all the time to decrypt restic repo
passwords. The command I use is probably different from somebody else's.
So attempt to make that and the inputs flexible.

Replaces type that generates the restic flag "--password-file". It
would be better to produce that flag separately because it's just a
string value doesn't require much logic. However, the password-command
flag could be anything, so some specialized configuration logic seems
necessary here.

In the near future, am planning to add a configuration type for restic
default flags. There could be a PasswordFile field that feeds into the
--password-file flag.