refactor(cmd): Use github.com/urfave/cli/v2
Would like there to be less boilerplate in package cmd.

Rename the restic subcommand to exec so that it matches the referenced
package. It's also good that the new name doesn't rhyme with the thing
it's wrapping. This should lead to clearer communication and definitely
less typos in the long run.

Also remove the flag -c from that subcommand and replace it with a
subcommand, which by the way, wraps a subcommand.