refactor(config): Simplify merge functionality
Removes a lot of functions that merge either the password config, or
restic types. These are artifacts of before the time I had started using
github.com/imdario/mergo to do a lot of the configuration merging. Now
it should be easier to add newer functionality, such as types to support
other restic subcommands.